backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
langchain==0.0.267
openai==0.27.8
pinecone-client==2.2.1
pydantic==1.10.8
python-jose==3.3.0
passlib==1.7.4
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
numpy==1.24.3
python-multipart==0.0.6
faiss-cpu==1.7.4
sentence-transformers==2.2.2